move class to separate folder
