+++ src/components/blocks/SetPointSimple/default.vue
@@ -1,0 +1,14 @@
+<template>
+  <q-card style="margin-bottom: 20px">
+    <q-card-title>SetPointSimple ({{ id }})</q-card-title>
+
+    <q-card-main>
+      <p class="text-weight-medium">Settings</p>
+      <div>
+        <span>Value</span><span>{{ settings.value }}</span>
+      </div>
+    </q-card-main>
+  </q-card>
+</template>
+
+<script lang="ts" src="./SetPointSimple.ts"></script>
